# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Xya-StyleMind Backend API
  version: 0.1.0
  description: |
    OpenAPI contract for the Xya-StyleMind Go/Gin backend.

    All JSON API endpoints use the shared response shape:
    `{ success, message, data, meta }`.

    Authentication uses Bearer JWT access tokens. Admin endpoints require an
    authenticated user with role `admin`.
servers:
  - url: http://localhost:8080
    description: Local backend
  - url: http://localhost:8080/api/v1
    description: Local API v1 base path
...
security: []
paths:
...
  /api/v1/me/notifications:
    get:
      tags: [Notifications]
      summary: List current user's notifications
      operationId: listMyNotifications
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: unread
          in: query
          description: When true, only unread notifications are returned.
          schema:
            type: boolean
      responses:
        '200':
          description: Notification list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/me/notifications/{id}/read:
    patch:
      tags: [Notifications]
      summary: Mark one notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/me/notifications/read-all:
    patch:
      tags: [Notifications]
      summary: Mark all current user's notifications as read
      operationId: markAllNotificationsRead
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notifications marked read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkAllNotificationsReadResponseEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
  responses:
    BadRequest:
      description: Invalid payload, validation error, or invalid identifier/filter.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: validation failed
    Unauthorized:
      description: Missing, invalid, expired, or revoked Bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: unauthorized
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: not found
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: too many requests
    InternalServerError:
      description: Unexpected server error. Response does not expose internal SQL/stack details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              value:
                success: false
                message: internal server error
...
  schemas:
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: validation failed
    PaginationMeta:
      type: object
      required: [page, limit, total, total_page, total_pages]
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          example: 42
        total_page:
          type: integer
          format: int64
          description: Backward-compatible pagination field.
          example: 3
        total_pages:
          type: integer
          format: int64
          example: 3
...
    Notification:
      type: object
      description: User-facing notification. Metadata is sanitized and must not contain tokens, passwords, secrets, or internal errors.
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        type:
          type: string
          example: order.status_updated
        title:
          type: string
        message:
          type: string
        metadata:
          type: object
          additionalProperties: true
          example:
            order_id: 3f9b5f41-7976-49a1-a801-fd4e339b75b5
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    NotificationResponseEnvelope:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Notification'
    NotificationListResponseEnvelope:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    MarkAllNotificationsReadResponse:
      type: object
      properties:
        updated:
          type: integer
          format: int64
    MarkAllNotificationsReadResponseEnvelope:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/MarkAllNotificationsReadResponse'
...